1. 부활 수정 - ResourceManager의 Clear 함수와 SceneChange의 순서로 인해 Scene -> Init 시 리소스를 불러오지 못하는 버그 수정 -> Clear 함수 발동 시점이 Scene의 소멸자가 아닌 SceneChange 함수 내에서로 변경 - 부활 시 MyPlayer 및 크리쳐가 정상적으로 생성되지 않는 버그 수정 - 부활 시 UI가 노출되지 않는 버그 수정
2. 새로운 몬스터 - 모블린 추가
- 뱀의 2배 공격력을 가짐
- 플레이어에게 4대 맞으면 쓰러짐

3. 버그 수정
- 사운드 버퍼가 정상적으로 Release 되지 않는 버그 수정
-> 스마트 포인터가 아닌 원시 포인터를 사용하되 적절하게 delete함.
- 구매 창이 눌리지 않는 버그 수정
- 구매 창이 열린 상태에서 드래그로 아이템을 팔면 크래시가 발생하는 버그 수정
- 부활 시 퀘스트 목표 개수가 0으로 보이는 버그 수정

+++ Server/Arrow.h
@@ -20,7 +20,7 @@ public:
 	bool IsHit() { return _hit; }
 
 protected:
-	uint64 _waitUntil = 0;
+	float _waitUntil = 0;
 	MonsterRef _target;
 
 	// 삭제를 위한 적중 여부
